requirements.txt
openai==1.12.0
python-dotenv==1.0.0
numpy==1.24.3
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pandas==2.1.0